Refactored home page.
