+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-portfolio",
   "version": "0.1.0",
-  "homepage": "https://t3mpz.github.io/React-Portfolio/",
+  "homepage": "/",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.10.5",
