+++ src/App.js
@@ -1,4 +1,3 @@
-import { useEffect } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
 import Navigation from "./components/Navigation";
@@ -8,13 +7,8 @@ import AboutMePage from "./pages/AboutMePage";
 import ProjectCardsPage from "./pages/ProjectCardsPage";
 import ResumePage from "./pages/ResumePage";
 import ContactPage from "./pages/ContactPage";
-// import Skills from "./components/Skills";
-// import Projects from "./components/Projects";
-// import Resume from "./components/Resume";
-// import About from "./components/About";
 
 function App() {
- 
   return (
     <>
       <Router>
@@ -25,11 +19,6 @@ function App() {
           <Route path="/project-cards" element={<ProjectCardsPage />}></Route>
           <Route path="/resume" element={<ResumePage />}></Route>
           <Route path="/contact" element={<ContactPage />}></Route>
-
-          {/* <Route path="/React-Portfolio" element={<About />}></Route>
-          <Route path="/skills" element={<Skills />}></Route>
-          <Route path="/projects" element={<Projects />}></Route>
-          <Route path="/resume" element={<Resume />}></Route> */}
         </Routes>
         <Footer />
       </Router>
