+++ src/pages/HomePage.jsx
@@ -1,12 +1,16 @@
 import { Box, Button, Container } from "@mui/material";
-import React, { useCallback } from "react";
-import Home from "../components/HomePage";
+import React, { useCallback, useEffect } from "react";
+import Home from "../components/Home";
 import { AiOutlineArrowRight } from "react-icons/ai";
 import { useNavigate } from "react-router-dom";
+import Socials from "../components/Socials";
 
 export default function HomePage() {
   const navigate = useNavigate();
   const handleOnClick = useCallback(() => navigate("/about-me", { replace: true }), [navigate]);
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
   return (
     <Container sx={{ minHeight: "85vh", mt: 5, textAlign: "center" }}>
       <Home />
@@ -15,6 +19,7 @@ export default function HomePage() {
           See more about me <AiOutlineArrowRight />
         </Button>
       </Box>
+      <Socials />
     </Container>
   );
 }
